+++ src/rpc/server.h
@@ -34,6 +34,8 @@ namespace barch {
             virtual ~rpc() = default;
             virtual call_result call(heap::vector<Variable>& result, const std::vector<std::string>& params) = 0;
 
+            virtual call_result call(heap::vector<Variable>& result, const heap::vector<std::string>& params) = 0;
+
             virtual call_result call(heap::vector<Variable>& result, const std::vector<std::string_view>& params) = 0;
 
             virtual call_result call(heap::vector<Variable>& result, const heap::vector<art::value_type>& params) = 0;